requirements.txt
flask==2.3.2
transformers==4.34.1
pytesseract==0.3.10
Pillow==10.0.0
SpeechRecognition==3.10.0
python-dotenv==1.0.0